+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-3.1.2 (unreleased)
+3.2.0 (2021-06-08)
 ------------------
 
 - Add new behavior "design.plone.contenttypes.behavior.show_modified".